move relevant code to models.py
